Agregar multiple usuarios

+++ src/main/java/com/portfolio/miportfolio/controller/EstudiosRestController.java
@@ -1,10 +1,11 @@
 package com.portfolio.miportfolio.controller;
 
 import com.portfolio.miportfolio.entity.Estudio;
-import com.portfolio.miportfolio.entity.Persona;
 import com.portfolio.miportfolio.service.IEstudiosService;
+import com.portfolio.miportfolio.service.UsuarioService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,9 +16,12 @@ public class EstudiosRestController {
     @Autowired
     private IEstudiosService estudiosService;
 
+    @Autowired
+    private UsuarioService usuarioService;
+
     @GetMapping("/estudios")
-    public List<Estudio> index() {
-        return estudiosService.findAll();
+    public List<Estudio> index(@RequestParam(name = "id_persona", required = true) String idPersona) {
+        return estudiosService.findByIdPersona(Long.parseLong(idPersona));
     }
 
     @GetMapping("/estudios/{id}")
@@ -34,8 +38,12 @@ public class EstudiosRestController {
     @PutMapping("/estudios/{id}")
     @ResponseStatus(HttpStatus.CREATED)
     public Estudio update(@RequestBody Estudio estudio, @PathVariable Long id) {
-
         Estudio estudioActual = estudiosService.findById(id);
+
+        var usuarioActual = this.usuarioService.getUsuarioLogueado();
+        if (usuarioActual == null || ! usuarioActual.getId().equals(estudioActual.getPersona().getUsuario().getId())) {
+            throw new AccessDeniedException("Acción no permitida");
+        }
 
         estudioActual.setCiudad(estudio.getCiudad());
         estudioActual.setDireccion(estudio.getDireccion());
@@ -53,6 +61,13 @@ public class EstudiosRestController {
     @DeleteMapping("/estudios/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Long id) {
+        Estudio estudioAEliminar = estudiosService.findById(id);
+
+        var usuarioActual = this.usuarioService.getUsuarioLogueado();
+        if (usuarioActual == null || ! usuarioActual.getId().equals(estudioAEliminar.getPersona().getUsuario().getId())) {
+            throw new AccessDeniedException("Acción no permitida");
+        }
+
         estudiosService.delete(id);
     }
 }
