+++ src/main/java/com/portfolio/miportfolio/controller/HardSkillRestController.java
@@ -2,21 +2,28 @@ package com.portfolio.miportfolio.controller;
 
 import com.portfolio.miportfolio.entity.Estudio;
 import com.portfolio.miportfolio.entity.HardSkill;
+import com.portfolio.miportfolio.service.IEstudiosService;
 import com.portfolio.miportfolio.service.IHardSkillService;
+import com.portfolio.miportfolio.service.UsuarioService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 @RestController
 @RequestMapping("/api")
 public class HardSkillRestController {
+
+    @Autowired
+    private UsuarioService usuarioService;
+
     @Autowired
     private IHardSkillService hardSkillService;
 
     @GetMapping("/hardskills")
-    public List<HardSkill> index() {
-        return hardSkillService.findAll();
+    public List<HardSkill> index(@RequestParam(name = "id_persona", required = true) String idPersona) {
+        return hardSkillService.findByIdPersona(Long.parseLong(idPersona));
     }
 
     @GetMapping("/hardskills/{id}")
@@ -36,6 +43,11 @@ public class HardSkillRestController {
 
         HardSkill hardSkill1Actual = hardSkillService.findById(id);
 
+        var usuarioActual = this.usuarioService.getUsuarioLogueado();
+        if (usuarioActual == null || ! usuarioActual.getId().equals(hardSkill1Actual.getPersona().getUsuario().getId())) {
+            throw new AccessDeniedException("Acción no permitida");
+        }
+
         hardSkill1Actual.setNombre(hardSkill.getNombre());
         hardSkill1Actual.setPorcentaje(hardSkill.getPorcentaje());
 
@@ -45,6 +57,12 @@ public class HardSkillRestController {
     @DeleteMapping("/hardskills/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Long id) {
+        HardSkill hardSkillAEliminar = hardSkillService.findById(id);
+
+        var usuarioActual = this.usuarioService.getUsuarioLogueado();
+        if (usuarioActual == null || ! usuarioActual.getId().equals(hardSkillAEliminar.getPersona().getUsuario().getId())) {
+            throw new AccessDeniedException("Acción no permitida");
+        }
         hardSkillService.delete(id);
     }
 }
